+++ contracts/Core/storage/BlockStorage.sol
@@ -6,14 +6,14 @@ contract BlockStorage {
     //epoch -> address -> dispute
     mapping(uint32 => mapping(address => Structs.Dispute)) public disputes;
     //epoch -> blockId -> block
-    mapping(uint32 => mapping(uint16 => Structs.Block)) public proposedBlocks;
+    mapping(uint32 => mapping(uint32 => Structs.Block)) public proposedBlocks;
     //epoch->blockId
-    mapping(uint32 => uint16[]) public sortedProposedBlockIds;
+    mapping(uint32 => uint32[]) public sortedProposedBlockIds;
     //stakerId->epoch
     mapping(uint32 => uint32) public epochLastProposed;
 
     // slither-disable-next-line constable-states
-    uint16 public numProposedBlocks;
+    uint32 public numProposedBlocks;
     // slither-disable-next-line constable-states
     int8 public blockIndexToBeConfirmed; // Index in sortedProposedBlockIds
     // epoch -> blocks
